Enscapsulate player class members
Turned player into partial class, and move all of the state classes as a nested class of the Knight partial class.

This way, all of the state classes are able to access and manipulate the private variables of Knight (just like a friend class in c++)

After setting this up, I was able to make all of the class members and data private. This way, classes other than the knight and it's states will not know about it's inner implementation details.